update to use docutils
