+++ docs/docutils.py
@@ -2,9 +2,11 @@
 Utilities for processing docs (notebooks mostly)
 """
 
+import os
+import sys
 import sciris as sc
-import nbconvert
 import nbformat
+import nbconvert.preprocessors as nbp
 
 default_folders = ['tutorials', 'user_guide'] # Folders with Jupyter notebooks
 temp_patterns = ['**/my-*.*', '**/example*.*'] # Temporary files to be removed
@@ -91,41 +93,74 @@ def init_cache(folders=None):
 
 def execute_notebook(path):
     """ Executes a single Jupyter notebook and returns success/failure """
-    try:
-        with open(path) as f:
-            with sc.timer(verbose=False) as T:
-                name = path.name
-                print(f'Executing {name}...')
+    with sc.timer(label=f'Execution time for {path}') as T:
+        try:
+            with open(path) as f:
+                print(f'Executing {path}...')
                 nb = nbformat.read(f, as_version=4)
-                ep = nbp.ExecutePreprocessor(timeout=timeout)
+                ep = nbp.ExecutePreprocessor(timeout=timeout, kernel_name='python3')
                 ep.preprocess(nb, {'metadata': {'path': os.path.dirname(path)}})
-        out = f'{yay} {name} executed successfully after {T.total:0.1f} s.'
-        print(out)
-        return out
-    except nbp.CellExecutionError as e:
-        return f'{boo} Execution failed for {path}: {str(e)}'
-    except Exception as e:
-        return f'{boo} Error processing {path}: {str(e)}'
+            string = f'{yay} {path} executed successfully '
+        except nbp.CellExecutionError as e:
+            string = f'{boo} Execution failed for {path}: {str(e)}\n'
+        except Exception as e:
+            string = f'{boo} Error processing {path}: {str(e)}\n'
+    string += f'(time: {T.total:0.1f} s)'
+    return string
+
 
 
 @sc.timer('Executed notebooks')
-def execute_notebooks(folders=None):
-    """ Executes the notebooks in parallel and prints the results """
-    notebooks = get_filenames(folders=folders)
-    results = sc.parallelize(execute_notebook, notebooks)
-    string = sc.newlinejoin(results)
-    
+def execute_notebooks(*args, folders=None):
+    """Executes the notebooks in parallel and prints the results.
+
+    Uses sys.argv[1:] if provided; otherwise uses folders to find notebooks.
+    """
+    T = sc.timer()
+    cwd = sc.thispath(__file__)
+    results = sc.objdict()
+    string = ''
+
+    # Determine which notebooks to run
+    if args:
+        notebooks = [sc.path(notebook).resolve() for notebook in args]
+    else:
+        notebooks = []
+        for folder in folders:
+            folder_path = cwd / folder
+            notebooks += [folder_path / f for f in sc.getfilepaths(folder_path, '*.ipynb')]
+
+    # Wrapper to chdir before execution
+    def execute_with_chdir(path):
+        os.chdir(path.parent)
+        return execute(path.name)
+
+    sc.heading(f'Running {len(notebooks)} notebooks...')
+    out = sc.parallelize(execute_with_chdir, notebooks, maxcpu=0.9, interval=0.3)
+    string += sc.strjoin(out, sep=f'\n\n\n{"—"*90}\n')
+    for nb, res in zip(notebooks, out):
+        results[str(nb)] = res
+
     sc.heading('Results')
     print(string)
     
     sc.heading('Summary')
     n_yay = string.count(yay)
     n_boo = string.count(boo)
-    summary = f'{n_yay} succeeded, {n_boo} failed'
-    if n_boo:
-        for i in range(len(notebooks)):
-            if boo in results[i]:
-                summary += f'\nFailed: {notebooks[i]}'
-    print(summary + '\n')
+    summary = f'{n_yay} succeeded, {n_boo} failed\n'
+
+    results.sort('values')
+    for nb,res in results.items():
+        timestr = res.split('time: ')[1][:-1]
+        suffix = f'{sc.path(nb).name:30s} ({timestr})'
+        if yay in res:
+            summary += f'\n{sc.ansi.green("Succeeded")}: {suffix}'
+        elif boo in res:
+            summary += f'\n{sc.ansi.red("   Failed")}: {suffix}'
+        else:
+            print('Unexpected result for {nb}, neither succeeded nor failed!')
+    print(summary)
+
+    T.toc()
     
     return results
