+++ docs/tutorials/check_tutorials.py
@@ -3,13 +3,13 @@
 Run all the Jupyter notebook tutorials. Usage:
   ./check_tutorials.py
 
-See check_notebooks in the parent folder for more information.
+See docutils.py in the parent folder for more information.
 """
 import os
 import sys
 import sciris as sc
 parent = sc.path(__file__).resolve().parent.parent # Ensure it's the real path if it's symlinked
-cnb = sc.importbypath(parent / 'check_notebooks.py')
+ut = sc.importbypath(parent / 'docutils.py')
 
 if __name__ == '__main__':
-    results = cnb.main(*sys.argv[1:], folders=['tutorials'])
+    results = ut.execute_notebooks(*sys.argv[1:], folders=['tutorials'])
